requirements.txt
pillow
boto3
opencv-python-headless
numpy
# torch